Factor out some dropwizard helper functions
This also makes the formerly java-specific one language agnostic.
